requirements.txt
Pillow==10.1.0
pillow-avif-plugin==1.5.0
